Add dependency injection using hilt
