+++ app/build.gradle
@@ -52,6 +52,9 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 
+    // Activity Ktx
+    implementation "androidx.activity:activity-ktx:1.2.0"
+
     //Rxjava
     implementation 'io.reactivex.rxjava2:rxjava:2.2.2'
     implementation 'io.reactivex.rxjava2:rxandroid:2.1.0'
@@ -69,9 +72,11 @@ dependencies {
     implementation("androidx.viewpager2:viewpager2:1.0.0")
 
     //Hilt - Dagger2 DI - both included in this
-    implementation("com.google.dagger:hilt-android:2.29.1-alpha")
-    kapt("com.google.dagger:hilt-android-compiler:2.29.1-alpha")
-    implementation("androidx.hilt:hilt-lifecycle-viewmodel:1.0.0-alpha02")
+    implementation "com.google.dagger:hilt-android:$hilt_version"
+    kapt "com.google.dagger:hilt-android-compiler:$hilt_version"
+    implementation "androidx.hilt:hilt-lifecycle-viewmodel:1.0.0-alpha03"
+    kapt "androidx.hilt:hilt-compiler:1.0.0-alpha03"
+
     //Hilt Test
     testImplementation("com.google.dagger:hilt-android-testing:2.29.1-alpha")
     kaptTest("com.google.dagger:hilt-android-compiler:2.29.1-alpha")
@@ -93,9 +98,7 @@ dependencies {
     // For testing
     androidTestImplementation "androidx.arch.core:core-testing:2.1.0"
 
-    def fragment_ktx = "1.2.5"
-    implementation "androidx.fragment:fragment-ktx:$fragment_ktx"
-
+    // Lifecycle
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.2.0"
 
 }
