+++ app/src/main/java/com/example/minicheckout/checkout/data/models/ExampleViewModel.kt
@@ -1,18 +1,15 @@
 package com.example.minicheckout.checkout.data.models
 
-import androidx.hilt.lifecycle.HiltViewModelFactory
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.minicheckout.repository.Repository
+import com.example.minicheckout.repository.CheckoutRepository
 import com.example.minicheckout.repository.network.data.BoxResponse
-import javax.inject.Inject
 
-class ExampleViewModel @ViewModelInject constructor() : ViewModel() {
-
-    @Inject
-    lateinit var repository: Repository
+class ExampleViewModel @ViewModelInject constructor(
+    private val repository: CheckoutRepository) :
+    ViewModel() {
 
     private var boxResponseLiveData = MutableLiveData<BoxResponse>()
