+++ app/src/main/java/com/example/minicheckout/core/dependencyinjection/AppModule.kt
@@ -3,16 +3,17 @@ package com.example.minicheckout.core.dependencyinjection
 import android.app.Application
 import com.example.minicheckout.core.coroutines.AppCoroutineContextProvider
 import com.example.minicheckout.core.coroutines.CoroutineContextProvider
-import com.example.minicheckout.repository.network.api.TestAPI
 import com.example.minicheckout.main.MainApplication
+import com.example.minicheckout.repository.network.api.APIService
+import com.example.minicheckout.repository.network.api.TestAPI
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ApplicationComponent
+import dagger.hilt.components.SingletonComponent
 
 //Dagger
 @Module
-@InstallIn(ApplicationComponent::class)
+@InstallIn(SingletonComponent::class)
 object AppModule {
 
     @Provides
@@ -22,5 +23,5 @@ object AppModule {
     fun contextProvider(): CoroutineContextProvider = AppCoroutineContextProvider()
 
     @Provides
-    fun provideAPIService() = TestAPI.apiService
+    fun provideAPIService(): APIService = TestAPI.apiService
 }
