+++ app/src/main/java/com/example/minicheckout/core/dependencyinjection/RepositoryModule.kt
@@ -1,21 +1,19 @@
 package com.example.minicheckout.core.dependencyinjection
 
+import com.example.minicheckout.repository.CheckoutRepository
 import com.example.minicheckout.repository.Repository
 import com.example.minicheckout.repository.network.api.APIService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ApplicationComponent
-import javax.inject.Singleton
+import dagger.hilt.android.components.ViewModelComponent
+import dagger.hilt.android.scopes.ViewModelScoped
 
 @Module
-@InstallIn(ApplicationComponent::class)
+@InstallIn(ViewModelComponent::class)
 object RepositoryModule {
 
     @Provides
-    @Singleton
-    fun providesRepository(apiService: APIService): Repository {
-        return Repository(apiService)
-    }
-
+    @ViewModelScoped
+    fun providesRepository(apiService: APIService): Repository = CheckoutRepository(apiService)
 }
