+++ app/src/main/java/com/example/minicheckout/repository/network/api/APIService.kt
@@ -5,7 +5,6 @@ import io.reactivex.Single
 import retrofit2.http.GET
 
 interface APIService {
-    //Function to get the API response
     @GET("api/current_fix")
-    fun currentFix(): Single<BoxResponse>
+    fun getCurrentFix(): Single<BoxResponse>
 }
