+++ build.gradle
@@ -1,6 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
     ext.kotlin_version = "1.3.72"
+    ext.hilt_version = "2.29.1-alpha"
     repositories {
         google()
         jcenter()
@@ -8,7 +9,7 @@ buildscript {
     dependencies {
         classpath "com.android.tools.build:gradle:4.1.2"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath("com.google.dagger:hilt-android-gradle-plugin:2.29.1-alpha")
+        classpath("com.google.dagger:hilt-android-gradle-plugin:$hilt_version")
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
